// orbito-frontend/src/core/api/generated/models/createSubscriptionCommand.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Orbito API
 * OpenAPI spec version: v1
 */

export interface CreateSubscriptionCommand {
  clientId?: string;
  planId?: string;
  amount?: number;
  /** @nullable */
  currency?: string | null;
  billingPeriodValue?: number;
  /** @nullable */
  billingPeriodType?: string | null;
  trialDays?: number;
}